+++ src/AdventOfCode/2022/Day_09.cs
@@ -10,7 +10,7 @@ public class Day_09
     [Example(answer: 1, "R 4;U 4;L 3;D 1;R 4;D 1;L 5;R 2")]
     [Example(answer: 36, "R 5;U 8;L 8;D 3;R 17;D 10;L 25;U 20")]
     [Puzzle(answer: 2619, O.μs100)]
-    public long part_two(Lines lines) => Simulate(lines, 10);
+    public int part_two(Lines lines) => Simulate(lines, 10);
     
     static int Simulate(Lines lines, int size)
     {
